Se agregó Swagger para documentación

+++ patitas-service/src/main/java/com/disenio/dao/PersonaDAO.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository("PersonaDAO")
-public interface PersonaDAO extends JpaRepository<Persona,Long> {
+public interface PersonaDAO extends JpaRepository<Persona,Integer> {
 }
